+++ app/smartbrain_3000/web/_app/immutable/nodes/13.Cf10gD0M.js
@@ -1,1 +1,0 @@
-import{nt as e,o as t,r as n,rt as r}from"../chunks/Ch8MiCKH.js";import{t as i}from"../chunks/CZcLIMKG.js";import"../chunks/xihTtKlq.js";import"../chunks/BOlQ_C0M.js";import"../chunks/uIBEvwzc.js";import{t as a}from"../chunks/DwMxFkqe.js";function o(o,s){r(s,!1),n(()=>i(`/settings/providers`)),t(),a(o,{block:!0}),e()}export{o as component};